De-branding pass 1

+++ build.ps1
@@ -1,12 +1,12 @@
 Param(
-    [Switch]$bootstrap = $false,
     [Switch]$dns = $false,
     [Switch]$dhcp = $false,
     [Switch]$tftp = $false,
-    [Switch]$http = $false
+    [Switch]$http = $false,
+    [Switch]$noshell = $false
 )
 
-Write-Host "Welcome to the Tenagra.net DevOps Build Environment."
+Write-Host "Welcome to the DevOps Build Environment."
 
 # Prepare environment
 if (Test-Path './.env')
@@ -22,28 +22,43 @@ if (Test-Path './.env')
 # Prepare environment
 $profiles = @()
 $TR
-if ($bootstrap -or $dns -or ("$env:TDK_DNS".ToUpper() -in ('TRUE', '1', 'Y', 'YES')))
+if ($dns -or ("$env:TDK_DNS".ToUpper() -in ('TRUE', '1', 'Y', 'YES')))
 {
     $profiles += '--profile dnsserver'
 }
-if ($bootstrap -or $dhcp -or ("$env:TDK_DHCP".ToUpper() -in ('TRUE', '1', 'Y', 'YES')))
+if ($dhcp -or ("$env:TDK_DHCP".ToUpper() -in ('TRUE', '1', 'Y', 'YES')))
 {
     $profiles += '--profile dhcpserver'
 }
-if ($bootstrap -or $tftp -or ("$env:TDK_TFTP".ToUpper() -in ('true', '1', 'Y', 'YES')))
+if ($tftp -or ("$env:TDK_TFTP".ToUpper() -in ('true', '1', 'Y', 'YES')))
 {
     $profiles += '--profile tftpserver'
 }
-if ($bootstrap -or $http -or ("$env:TDK_http".ToUpper() -in ('true', '1', 'Y', 'YES')))
+if ($http -or ("$env:TDK_http".ToUpper() -in ('true', '1', 'Y', 'YES')))
 {
     $profiles += '--profile webserver'
 }
 
 # Run the Toolkit container
-Push-Location ./services
-Register-EngineEvent PowerShell.Exiting -Action { Pop-Location } | Out-Null
-
-Invoke-Expression "docker compose build"
-Invoke-Expression "docker compose $( $profiles -join ' ' ) up -d"
-Invoke-Expression "docker compose run toolkit /bin/bash"
-Invoke-Expression "docker compose $( $profiles -join ' ' ) down"
+try
+{
+    Push-Location ./services
+    Invoke-Expression "docker compose $( $profiles -join ' ' ) build"
+    if ($( Invoke-Expression "docker compose ls -q" ) -contains 'devopstdk')
+    {
+        Invoke-Expression "docker compose $( $profiles -join ' ' ) restart" | Out-Null
+    }
+    Invoke-Expression "docker compose $( $profiles -join ' ' ) up -d"
+    if (!$noshell)
+    {
+        Invoke-Expression "docker compose run toolkit /bin/bash"
+    }
+}
+finally
+{
+    if (!$noshell)
+    {
+        Invoke-Expression "docker compose $( $profiles -join ' ' ) down"
+    }
+    Pop-Location
+}
